chore: refactor facet fetch logic
